+++ analyseCoverage.mjs
@@ -1,3 +1,5 @@
+// @ts-check
+
 import fs from "fs";
 import { join } from "path";
 import { fileURLToPath } from "url";
@@ -5,16 +7,11 @@ import v8Coverage from "@bcoe/v8-coverage";
 import sourceRange from "./sourceRange.mjs";
 
 /**
- * Analyzes [Node.js generated V8 JavaScript code coverage data](https://nodejs.org/api/cli.html#cli_node_v8_coverage_dir)
+ * Analyzes
+ * [Node.js generated V8 JavaScript code coverage data](https://nodejs.org/api/cli.html#cli_node_v8_coverage_dir)
  * in a directory; useful for reporting.
- * @kind function
- * @name analyseCoverage
  * @param {string} coverageDirPath Code coverage data directory path.
  * @returns {Promise<CoverageAnalysis>} Resolves the coverage analysis.
- * @example <caption>How to import.</caption>
- * ```js
- * import analyseCoverage from "coverage-node/analyseCoverage.mjs";
- * ```
  */
 export default async function analyseCoverage(coverageDirPath) {
   if (typeof coverageDirPath !== "string")
@@ -29,6 +26,7 @@ export default async function analyseCoverage(coverageDirPath) {
         fs.promises
           .readFile(join(coverageDirPath, fileName), "utf8")
           .then((coverageFileJson) => {
+            /** @type {import("@bcoe/v8-coverage").ProcessCov} */
             const { result } = JSON.parse(coverageFileJson);
             return {
               // For performance, filtering happens as early as possible.
@@ -53,18 +51,11 @@ export default async function analyseCoverage(coverageDirPath) {
     await Promise.all(filteredProcessCoverages)
   );
 
-  // The analysis will only include info useful for reporting.
+  /** @type {CoverageAnalysis} */
   const analysis = {
-    // Total number of files.
     filesCount: 0,
-
-    // Fully covered file paths.
     covered: [],
-
-    // File paths and ignored ranges.
     ignored: [],
-
-    // File paths and uncovered ranges.
     uncovered: [],
   };
 
@@ -83,17 +74,17 @@ export default async function analyseCoverage(coverageDirPath) {
       const uncovered = [];
 
       for (const range of uncoveredRanges) {
-        const { ignore, ...rangeDetails } = sourceRange(
+        const sourceCodeRange = sourceRange(
           source,
           range.startOffset,
           // The coverage data end offset is the first character after the
-          // range. For reporting to a user, it’s better to show the range
-          // as only the included characters.
+          // range. For reporting to a user, it’s better to show the range as
+          // only the included characters.
           range.endOffset - 1
         );
 
-        if (ignore) ignored.push(rangeDetails);
-        else uncovered.push(rangeDetails);
+        if (sourceCodeRange.ignore) ignored.push(sourceCodeRange);
+        else uncovered.push(sourceCodeRange);
       }
 
       if (ignored.length) analysis.ignored.push({ path, ranges: ignored });
@@ -104,3 +95,21 @@ export default async function analyseCoverage(coverageDirPath) {
 
   return analysis;
 }
+
+/**
+ * [Node.js generated V8 JavaScript code coverage data](https://nodejs.org/api/cli.html#cli_node_v8_coverage_dir)
+ * analysis; useful for reporting.
+ * @typedef {object} CoverageAnalysis
+ * @prop {number} filesCount Number of files analyzed.
+ * @prop {Array<string>} covered Covered file absolute paths.
+ * @prop {Array<SourceCodeRanges>} ignored Ignored source code ranges.
+ * @prop {Array<SourceCodeRanges>} uncovered Uncovered source code ranges.
+ */
+
+/**
+ * A source code file with ranges of interest.
+ * @typedef {object} SourceCodeRanges
+ * @prop {string} path File absolute path.
+ * @prop {Array<import("./sourceRange.mjs").SourceCodeRange>} ranges Ranges of
+ *   interest.
+ */