+++ coverage-node.mjs
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+// @ts-check
 
 import { spawn } from "child_process";
 import disposableDirectory from "disposable-directory";
@@ -15,10 +16,7 @@ import reportCoverage from "./reportCoverage.mjs";
  * Powers the `coverage-node` CLI. Runs Node.js with the given arguments and
  * coverage enabled. An analysis of the coverage is reported to the console, and
  * if coverage isn’t complete the process exits with code `1`.
- * @kind function
- * @name coverageNode
  * @returns {Promise<void>} Resolves when all work is complete.
- * @ignore
  */
 async function coverageNode() {
   try {
@@ -43,7 +41,7 @@ async function coverageNode() {
           const analysis = await analyseCoverage(tempDirPath);
           reportCoverage(analysis);
           if (analysis.uncovered.length) process.exitCode = 1;
-        } else process.exitCode = exitCode;
+        } else if (exitCode !== null) process.exitCode = exitCode;
       });
       // coverage ignore next line
     } else {
@@ -57,7 +55,7 @@ async function coverageNode() {
             `Skipped code coverage as Node.js is ${process.version}, v${minNodeVersion.major}.${minNodeVersion.minor}.${minNodeVersion.patch}+ is supported.`
           )}\n`
         );
-      else process.exitCode = exitCode;
+      else if (exitCode !== null) process.exitCode = exitCode;
     }
   } catch (error) {
     reportCliError(