+++ coverageSupported.mjs
@@ -1,3 +1,5 @@
+// @ts-check
+
 import coverageSupportedMinNodeVersion from "./coverageSupportedMinNodeVersion.mjs";
 import semver from "./semver.mjs";
 
@@ -9,15 +11,9 @@ const {
 const { major, minor, patch } = semver(process.versions.node);
 
 /**
- * Is the process Node.js version greater at least [the minimum required to
- * support code coverage]{@link coverageSupportedMinNodeVersion}.
- * @kind constant
- * @name coverageSupported
+ * Is the process Node.js version greater than or equal to
+ * {@link coverageSupportedMinNodeVersion the minimum required to support code coverage}.
  * @type {boolean}
- * @example <caption>How to import.</caption>
- * ```js
- * import coverageSupported from "coverage-node/coverageSupported.mjs";
- * ```
  */
 export default // coverage ignore next line
 (major === minMajor && minor >= minMinor && patch >= minPatch) ||