+++ sourceRange.mjs
@@ -1,14 +1,16 @@
+// @ts-check
+
 /**
  * Gets info about a source code range, including line/column numbers and if
  * it’s ignored by a comment.
- * @kind function
- * @name sourceRange
  * @param {string} source Source code.
  * @param {number} startOffset Start character offset.
  * @param {number} endOffset End character offset.
- * @param {string|boolean} [ignoreNextLineComment=" coverage ignore next line"] Single line case-insensitive comment content to ignore ranges that start on the the next line, or `false` to disable ignore comments.
+ * @param {string | false} [ignoreNextLineComment] Single line
+ *   case-insensitive comment content to ignore ranges that start on the the
+ *   next line, or `false` to disable ignore comments. Defaults to
+ *   `" coverage ignore next line"`.
  * @returns {SourceCodeRange} Source code range info.
- * @ignore
  */
 export default function sourceRange(
   source,
@@ -33,13 +35,25 @@ export default function sourceRange(
       "Fourth argument `ignoreNextLineComment` must be a string or `false`."
     );
 
-  if (ignoreNextLineComment)
-    var ignoreNextLineCommentLowerCase = `//${ignoreNextLineComment.toLowerCase()}`;
+  const ignoreNextLineCommentLowerCase = ignoreNextLineComment
+    ? `//${ignoreNextLineComment.toLowerCase()}`
+    : null;
 
-  const range = {
-    start: { offset: startOffset },
-    end: { offset: endOffset },
-  };
+  /** @type {SourceCodeRange["ignore"]} */
+  let ignore = false;
+
+  /** @type {SourceCodeLocation["line"] | undefined} */
+  let startLine;
+
+  /** @type {SourceCodeLocation["column"] | undefined} */
+  let startColumn;
+
+  /** @type {SourceCodeLocation["line"] | undefined} */
+  let endLine;
+
+  /** @type {SourceCodeLocation["column"] | undefined} */
+  let endColumn;
+
   const lines = source.split(/^/gmu);
 
   let lineOffset = 0;
@@ -48,33 +62,65 @@ export default function sourceRange(
     const nextLineOffset = lineOffset + lineSource.length;
 
     if (
-      !("line" in range.start) &&
+      !startLine &&
       startOffset >= lineOffset &&
       startOffset < nextLineOffset
     ) {
-      range.start.line = lineIndex + 1;
-      range.start.column = startOffset - lineOffset + 1;
-      range.ignore =
-        // Is ignoring enabled.
-        !!ignoreNextLineComment &&
-        // First line can’t be ignored.
-        !!lineIndex &&
-        // Does the previous line contain the case-insensitive comment to ignore
+      startLine = lineIndex + 1;
+      startColumn = startOffset - lineOffset + 1;
+
+      if (
+        // Ignoring is enabled.
+        ignoreNextLineCommentLowerCase &&
+        // It’s not the first line that can’t be ignored, because there can’t be
+        // an ignore comment on the previous line.
+        lineIndex &&
+        // The previous line contains the case-insensitive comment to ignore
         // this line.
         lines[lineIndex - 1]
           .trim()
           .toLowerCase()
-          .endsWith(ignoreNextLineCommentLowerCase);
+          .endsWith(ignoreNextLineCommentLowerCase)
+      )
+        ignore = true;
     }
 
     if (endOffset >= lineOffset && endOffset < nextLineOffset) {
-      range.end.line = lineIndex + 1;
-      range.end.column = endOffset - lineOffset + 1;
+      endLine = lineIndex + 1;
+      endColumn = endOffset - lineOffset + 1;
       break;
     }
 
     lineOffset = nextLineOffset;
   }
 
-  return range;
+  return {
+    ignore,
+    start: {
+      offset: startOffset,
+      line: /** @type {number} */ (startLine),
+      column: /** @type {number} */ (startColumn),
+    },
+    end: {
+      offset: endOffset,
+      line: /** @type {number} */ (endLine),
+      column: /** @type {number} */ (endColumn),
+    },
+  };
 }
+
+/**
+ * Source code location.
+ * @typedef {object} SourceCodeLocation
+ * @prop {number} offset Character offset.
+ * @prop {number} line Line number.
+ * @prop {number} column Column number.
+ */
+
+/**
+ * Source code range details.
+ * @typedef {object} SourceCodeRange
+ * @prop {boolean} ignore Should it be ignored.
+ * @prop {SourceCodeLocation} start Start location.
+ * @prop {SourceCodeLocation} end End location.
+ */