+++ test.mjs
@@ -1,4 +1,7 @@
+// @ts-check
+
 import TestDirector from "test-director";
+import test_CliError from "./CliError.test.mjs";
 import test_analyseCoverage from "./analyseCoverage.test.mjs";
 import test_childProcessPromise from "./childProcessPromise.test.mjs";
 import test_cli_coverage_node from "./coverage-node.test.mjs";
@@ -9,6 +12,7 @@ import test_replaceStackTraces from "./test/replaceStackTraces.test.mjs";
 
 const tests = new TestDirector();
 
+test_CliError(tests);
 test_analyseCoverage(tests);
 test_childProcessPromise(tests);
 test_cli_coverage_node(tests);